major ui changes.
completely changed how the inventory and interacting with it work. A
similar change for the combo lock should follow. there is an indicator
in the bottom right, inventory on the left, the text area top left and a
timer in the center. all of this needs tidying but should be good for
structural changes now